+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.1.1]
+### Added
+    * Proxy for local development when sending webmentions
+
 ## [0.1.0]
 ### Added
-  * The ability to send webmentions
-  * Configurable storage, defaulting to ets
+    * The ability to send webmentions
+    * Configurable storage, defaulting to ets